Revert "Merge pull request #11425 from comoyo/mozContacts2vcard"
This reverts commit 1b3b923534f606c69fce61136aa21849ee8bba6a, reversing
changes made to c354940fbe112bcf2d90e2a5654ad1824f3a2348.
