+++ apps/communications/contacts/test/unit/contacts_form_test.js
@@ -1,4 +1,3 @@
-require('/shared/test/unit/mocks/mock_contact_all_fields.js');
 //Avoiding lint checking the DOM file renaming it to .html
 requireApp('communications/contacts/test/unit/mock_form_dom.js.html');
 
@@ -6,6 +5,7 @@ requireApp('communications/contacts/js/contacts_form.js');
 requireApp('communications/contacts/js/utilities/templates.js');
 requireApp('communications/contacts/js/utilities/dom.js');
 requireApp('communications/contacts/test/unit/mock_contacts.js');
+requireApp('communications/contacts/test/unit/mock_contact_all_fields.js');
 requireApp('communications/contacts/test/unit/mock_fb.js');
 
 var subject,
